Restructured code files and pointed the py files through main in order to make merging easier.
